requirements.txt
fastapi
uvicorn
pandas
google-generativeai
sentence-transformers
python-dotenv